pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "t3"
version = "2024.5.30"
requires-python = ">= 3.8"
dependencies = [
    "torch == 2.1.0",
    "torchvision == 0.16",
    "numpy",
    "hydra-core",
    "tqdm",
    "wandb",
    "pandas",
    "pyarrow",
    "autolab_core",
    "scipy",
    "timm",
    "webdataset",
    "natsort",
    "autolab_core",
]

[tool.setuptools]
packages = ["t3"]